fix bug of contracts

+++ packages/loopring_v3/contracts/amm/libamm/AmmUpdateProcess.sol
@@ -9,21 +9,16 @@ import "./AmmData.sol";
 import "./AmmUtil.sol";
 import "../../lib/MathUint96.sol";
 
-
 /// @title AmmUpdateProcess
-library AmmUpdateProcess
-{
-    using MathUint96        for uint96;
+library AmmUpdateProcess {
+    using MathUint96 for uint96;
     using TransactionReader for ExchangeData.Block;
 
     function approveAmmUpdates(
-        AmmData.State      storage S,
-        AmmData.Context    memory  ctx,
-        bool                       start
-        )
-        internal
-        view
-    {
+        AmmData.State storage S,
+        AmmData.Context memory ctx,
+        bool start
+    ) internal view {
         // Start by reading the first 28 bytes into packedData
         uint txsDataPtr = ctx.txsDataPtr + 28;
         for (uint i = 0; i < ctx.tokens.length; i++) {
@@ -57,19 +52,31 @@ library AmmUpdateProcess
                 // update.accountID == ctx.accountID &&
                 // update.tokenID == token.tokenID &&
                 // update.feeBips == ctx.feeBips &&
-                packedDataA & 0xffffffffffffffffffffffffffffffffffffffffffffffffffffffff ==
-                (uint(ExchangeData.TransactionType.AMM_UPDATE) << 216) | (uint(address(this)) << 56) | (ctx.accountID << 24) | (token.tokenID << 8) | ctx.feeBips,
+                packedDataA &
+                    0xffffffffffffffffffffffffffffffffffffffffffffffffffffffff ==
+                    (uint(ExchangeData.TransactionType.AMM_UPDATE) << 216) |
+                        (uint(address(this)) << 56) |
+                        (ctx.accountID << 24) |
+                        (token.tokenID << 8) |
+                        ctx.feeBips,
                 "INVALID_AMM_UPDATE_TX_DATA"
             );
 
             if (start) {
                 // Bring actual balances and virtual balances from L2 to L1
-                ctx.tokenBalancesL2[i] = /*balance*/uint96(packedDataB & 0xffffffffffffffffffffffff);
-                ctx.vTokenBalancesL2[i] = /*vbalance*/uint96((packedDataB >> 128) & 0xffffffffffffffffffffffff);
+                ctx.tokenBalancesL2[i] = /*balance*/ uint96(
+                    packedDataB & 0xffffffffffffffffffffffff
+                );
+                ctx.vTokenBalancesL2[i] = /*vbalance*/ uint96(
+                    (packedDataB >> 128) & 0xffffffffffffffffffffffff
+                );
             } else {
                 // Verify new virtual balances are the same value as on L2.
                 require(
-                    ctx.vTokenBalancesL2[i] == /*vbalance*/uint96((packedDataB >> 128) & 0xffffffffffffffffffffffff),
+                    ctx.vTokenBalancesL2[i] ==
+                        /*vbalance*/ uint96(
+                            (packedDataB >> 128) & 0xffffffffffffffffffffffff
+                        ),
                     "INVALID_VIRTUAL_BALANCE_UPDATE"
                 );
             }
@@ -80,10 +87,14 @@ library AmmUpdateProcess
         if (start && ctx.settings.assetManager != IAssetManager(0)) {
             // Add the L1 balances on top of the balances on L2
             for (uint i = 0; i < ctx.tokens.length; i++) {
-                ctx.tokenBalancesL2[i] = ctx.tokenBalancesL2[i].add(S.balancesL1[ctx.tokens[i].addr]);
+                ctx.tokenBalancesL2[i] = ctx.tokenBalancesL2[i].add(
+                    S.balancesL1[ctx.tokens[i].addr]
+                );
             }
         }
 
-        ctx.txsDataPtr += ExchangeData.TX_DATA_AVAILABILITY_SIZE * ctx.tokens.length;
+        ctx.txsDataPtr +=
+            ExchangeData.TX_DATA_AVAILABILITY_SIZE *
+            ctx.tokens.length;
     }
 }
